requirements.txt
opencv-python
numpy
python-dotenv
insightface
pymongo
requests
faiss-cpu
watchdog
websockets
